Move tests for the ISCSI module
Move the unit tests for the ISCSI module to a new file.
