+++ tests/nosetests/pyanaconda_tests/module_storage_test.py
@@ -45,13 +45,6 @@ from pyanaconda.modules.common.task import TaskInterface
 from pyanaconda.modules.storage.fcoe import FCOEModule
 from pyanaconda.modules.storage.fcoe.discover import FCOEDiscoverTask
 from pyanaconda.modules.storage.fcoe.fcoe_interface import FCOEInterface
-from pyanaconda.modules.storage.iscsi import ISCSIModule
-from pyanaconda.modules.storage.iscsi.discover import ISCSIDiscoverTask, ISCSILoginTask
-from pyanaconda.modules.storage.iscsi.iscsi_interface import ISCSIInterface, \
-    ISCSIDiscoverTaskInterface
-from pyanaconda.modules.storage.constants import IscsiInterfacesMode
-from pyanaconda.modules.common.structures.iscsi import Portal, Credentials, Node
-from pyanaconda.modules.common.constants.objects import ISCSI
 from pyanaconda.modules.storage.installation import ActivateFilesystemsTask, \
     MountFilesystemsTask, WriteConfigurationTask
 from pyanaconda.modules.storage.partitioning.validate import StorageValidateTask
@@ -1403,140 +1396,6 @@ class StorageTasksTestCase(unittest.TestCase):
         patched_conf.target.is_directory = False
         WriteConfigurationTask(storage).run()
         write.assert_called_once_with(storage)
-
-
-class ISCSIInterfaceTestCase(unittest.TestCase):
-    """Test DBus interface of the iSCSI module."""
-
-    def setUp(self):
-        """Set up the module."""
-        self.iscsi_module = ISCSIModule()
-        self.iscsi_interface = ISCSIInterface(self.iscsi_module)
-
-        self._portal = Portal()
-        self._portal.ip_address = "10.43.136.67"
-        self._portal.port = "3260"
-
-        self._credentials = Credentials()
-        self._credentials.username = "mersault"
-        self._credentials.password = "nothing"
-        self._credentials.reverse_username = "tluasrem"
-        self._credentials.reverse_password = "thing"
-
-        self._node = Node()
-        self._node.name = "iqn.2014-08.com.example:t1"
-        self._node.address = "10.43.136.67"
-        self._node.port = "3260"
-        self._node.iface = "iface0"
-        self._node.net_ifacename = "ens3"
-
-        # Connect to the properties changed signal.
-        self.callback = Mock()
-        self.iscsi_interface.PropertiesChanged.connect(self.callback)
-
-    @patch('pyanaconda.modules.storage.iscsi.iscsi.iscsi')
-    def initator_property_test(self, iscsi):
-        """Test Initiator property."""
-        initiator_name = "iqn.1994-05.com.redhat:blabla"
-        iscsi.initiator_set = False
-        self.iscsi_interface.SetInitiator(initiator_name)
-        iscsi.initiator = initiator_name
-        self.assertEqual(self.iscsi_interface.Initiator, initiator_name)
-        iscsi.initiator_set = True
-        initiator_name2 = "iqn.1994-05.com.redhat:blablabla"
-        self.iscsi_interface.SetInitiator(initiator_name2)
-        self.callback.assert_called_once_with(ISCSI.interface_name, {'Initiator': initiator_name}, [])
-
-    @patch('pyanaconda.modules.storage.iscsi.iscsi.iscsi')
-    def can_set_initiator_test(self, iscsi):
-        """Test CanSetInitiator method."""
-        self.assertIsInstance(self.iscsi_interface.CanSetInitiator(), bool)
-
-    @patch('pyanaconda.modules.storage.iscsi.iscsi.iscsi')
-    def get_interface_mode_test(self, iscsi):
-        """Test GetInterfaceMode method."""
-        blivet_mode_values = ["none", "default", "bind"]
-        for blivet_mode in blivet_mode_values + ["unexpected_value"]:
-            iscsi.mode = blivet_mode
-            _mode = IscsiInterfacesMode(self.iscsi_interface.GetInterfaceMode())
-
-    @patch('pyanaconda.modules.storage.iscsi.iscsi.iscsi')
-    def is_node_from_ibft_test(self, iscsi):
-        """Test IsNodeFromIbft method."""
-        iscsi.ibft_nodes = []
-        result = self.iscsi_interface.IsNodeFromIbft(
-            self._unpack_structure_content(Node.to_structure(self._node))
-        )
-        self.assertFalse(result)
-
-        blivet_node = Mock()
-        blivet_node.name = self._node.name
-        blivet_node.address = self._node.address
-        blivet_node.port = int(self._node.port)
-        blivet_node.iface = self._node.iface
-        iscsi.ibft_nodes = [blivet_node]
-        result = self.iscsi_interface.IsNodeFromIbft(
-            self._unpack_structure_content(Node.to_structure(self._node))
-        )
-        self.assertTrue(result)
-
-    @patch('pyanaconda.modules.storage.iscsi.iscsi.iscsi')
-    def get_interface_test(self, iscsi):
-        """Test GetInterface method."""
-        iscsi.ifaces = {
-            "iface0" : "ens3",
-            "iface1" : "ens7",
-        }
-        self.assertEqual(self.iscsi_interface.GetInterface("iface0"), "ens3")
-        self.assertEqual(self.iscsi_interface.GetInterface("nonexisting"), "")
-
-    def _unpack_structure_content(self, structure):
-        return {key: value.unpack() for key, value in structure.items()}
-
-    @patch_dbus_publish_object
-    def discover_with_task_test(self, publisher):
-        """Test the discover task."""
-        interfaces_mode = "default"
-        task_path = self.iscsi_interface.DiscoverWithTask(
-            self._unpack_structure_content(Portal.to_structure(self._portal)),
-            self._unpack_structure_content(Credentials.to_structure(self._credentials)),
-            interfaces_mode
-        )
-
-        obj = check_task_creation(self, task_path, publisher, ISCSIDiscoverTask)
-
-        self.assertIsInstance(obj, ISCSIDiscoverTaskInterface)
-
-        self.assertEqual(obj.implementation._portal, self._portal)
-        self.assertEqual(obj.implementation._credentials, self._credentials)
-        self.assertEqual(obj.implementation._interfaces_mode, IscsiInterfacesMode.DEFAULT)
-
-    @patch_dbus_publish_object
-    def login_with_task_test(self, publisher):
-        """Test the login task."""
-        task_path = self.iscsi_interface.LoginWithTask(
-            self._unpack_structure_content(Portal.to_structure(self._portal)),
-            self._unpack_structure_content(Credentials.to_structure(self._credentials)),
-            self._unpack_structure_content(Node.to_structure(self._node)),
-        )
-
-        obj = check_task_creation(self, task_path, publisher, ISCSILoginTask)
-
-        self.assertEqual(obj.implementation._portal, self._portal)
-        self.assertEqual(obj.implementation._credentials, self._credentials)
-        self.assertEqual(obj.implementation._node, self._node)
-
-    @patch('pyanaconda.modules.storage.iscsi.iscsi.iscsi')
-    def reload_module_test(self, iscsi):
-        """Test ReloadModule."""
-        self.iscsi_interface.ReloadModule()
-        iscsi.startup.assert_called_once_with()
-
-    @patch('pyanaconda.modules.storage.iscsi.iscsi.iscsi')
-    def write_configuration_test(self, iscsi):
-        """Test WriteConfiguration."""
-        self.iscsi_interface.WriteConfiguration()
-        iscsi.write.assert_called_once_with(conf.target.system_root, None)
 
 
 class FCOEInterfaceTestCase(unittest.TestCase):
